Change header to follow new plan
Please note that this does not follow the plan as laid out in the project description, as the second summary is per person instead of per sentence. That is because checking differences per person actually seems to me like the more interesting and useful approach now that I thought of it more.